refactor: redo directory structure
- groups files related to renderable 'components'
