+++ src/components/mod.rs
@@ -1,0 +1,9 @@
+pub mod session_manager;
+
+pub use super::components::session_manager::*;
+
+mod popups;
+mod views;
+mod window;
+
+use super::components::{popups::*, views::*, window::*};
