+++ src/components/views/log/filter.rs
@@ -9,7 +9,7 @@ use serde::{Deserialize, Serialize};
 use strum::IntoEnumIterator;
 use strum_macros::{Display, EnumIter, FromRepr};
 
-use crate::task_timer::LogType;
+use crate::log_type::LogType;
 
 #[derive(Serialize, Deserialize, Default, EnumIter, Display, Clone, Copy, FromRepr, PartialEq)]
 pub enum Filter {
