+++ src/components/views/log/time_stamp.rs
