+++ src/components/views/paginator.rs
