+++ src/components/views/task_view/navigation_bar.rs
