+++ src/info_subtype.rs
