--- src/task_timer/mod.rs
@@ -1,14 +1,0 @@
-pub mod node;
-pub mod session_manager;
-
-mod info_subtype;
-mod log_type;
-mod popups;
-mod traits;
-mod views;
-mod window;
-
-use super::task_timer::{
-    info_subtype::InfoSubType, log_type::LogType, node::*, popups::PopupType, session_manager::SessionState,
-    traits::*, window::Window,
-};
